requirements-azure.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0
sqlalchemy==2.0.23
alembic==1.12.1
asyncpg==0.29.0
psycopg2-binary==2.9.9
aiosqlite==0.19.0
fastapi-users[sqlalchemy]==12.1.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.1.0
python-dotenv==1.0.0
pydantic-settings==2.1.0
httpx==0.25.2
aiohttp==3.9.1
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.2.0
pandas==2.1.4
numpy==1.26.2
python-dateutil==2.8.2
openpyxl==3.1.2
xlrd==2.0.1
